Refactor vacation requested field to Double and update services
Changed the type of the 'requested' field from Long to Double in Vacation-related DTOs and model to support fractional values. Updated validation constraints and adjusted related service logic to handle Double values. Added batch update and refresh logic for counters in CounterService and VacationGroupService, and extended repositories with new query methods.

+++ src/main/java/fr/antoinenebout/vacation_app_api/dto/Vacation/VacationCreateDTO.java
@@ -11,7 +11,7 @@ public class VacationCreateDTO {
 
     @NotNull(message = "Request value is required")
     @Min(value = 1, message = "Requested value must be greater than 1")
-    private Long requested;
+    private Double requested;
 
 
     @NotNull(message = "Vacation_type_id value is required")
